pyproject.toml
[project]
name = "agent-memory"
version = "2.0.0"
description = "A layered memory system for AI agents with vector search, knowledge graphs, and automatic memory evolution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client>=1.7.0",
    "openai>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
agent-memory-test = "agent_memory.memory_service:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_memory"]
